[ fix ] id 없음 문제 해결
